Consolidate backend modules
