--- backend/crawlers/example_crawler/settings.py
@@ -1,7 +1,0 @@
-BOT_NAME = "example_crawler"
-
-SPIDER_MODULES = ["example_crawler.spiders"]
-NEWSPIDER_MODULE = "example_crawler.spiders"
-
-ROBOTSTXT_OBEY = True
-FEED_EXPORT_ENCODING = "utf-8"
