--- backend/crawlers/example_crawler/spiders/example_spider.py
@@ -1,11 +1,0 @@
-import scrapy
-
-
-class ExampleSpider(scrapy.Spider):
-    name = "example"
-    allowed_domains = ["example.com"]
-    start_urls = ["https://example.com"]
-
-    def parse(self, response):
-        title = response.css("h1::text").get()
-        yield {"title": title, "url": response.url}
