--- backend/crawlers/scrapy.cfg
@@ -1,5 +1,0 @@
-[settings]
-default = example_crawler.settings
-
-[deploy]
-project = example_crawler
