+++ business_intel_scraper/backend/osint/integrations.py
@@ -14,11 +14,20 @@ environments where SpiderFoot or TheHarvester are not installed.
 
 from __future__ import annotations
 
+import json
 import shutil
 import subprocess
+from typing import Any
 
 
-def run_spiderfoot(domain: str) -> dict[str, str]:
+def _parse_json(text: str) -> Any:
+    try:
+        return json.loads(text)
+    except Exception:
+        return text.strip()
+
+
+def run_spiderfoot(domain: str, parse_output: bool = False) -> dict[str, Any]:
     """Run SpiderFoot against a domain.
 
     The function expects the ``spiderfoot`` (or ``sf.py``) executable to be
@@ -47,10 +56,12 @@ def run_spiderfoot(domain: str) -> dict[str, str]:
     cmd = [executable, "-q", domain, "-F", "json"]
     proc = subprocess.run(cmd, capture_output=True, text=True)
     output = proc.stdout.strip() or proc.stderr.strip()
+    if parse_output:
+        return {"domain": domain, "data": _parse_json(output)}
     return {"domain": domain, "output": output}
 
 
-def run_theharvester(domain: str) -> dict[str, str]:
+def run_theharvester(domain: str, parse_output: bool = False) -> dict[str, Any]:
     """Run TheHarvester against a domain.
 
     Similar to :func:`run_spiderfoot`, this wrapper relies on the presence of
@@ -76,4 +87,6 @@ def run_theharvester(domain: str) -> dict[str, str]:
     cmd = [executable, "-d", domain, "-b", "all"]
     proc = subprocess.run(cmd, capture_output=True, text=True)
     output = proc.stdout.strip() or proc.stderr.strip()
+    if parse_output:
+        return {"domain": domain, "data": _parse_json(output)}
     return {"domain": domain, "output": output}
